Add the function for showing the number of the operators and graphs

+++ test_ir.py
@@ -31,4 +31,4 @@ class Net(nn.Cell):
 context.set_context(save_graphs=True)
 net = Net()
 model = Model(net)
-model.predict(Tensor(np.ones((10, 10))))
+model.predict(Tensor(np.ones((10, 10), dtype=np.float32)))
